671 documentation and module organisation
Actions:

+ update and add documentation for translation traits and macros
+ create macro module for helper functions

Motivation:

+ up-to-date docs
+ de-clutter macro lib.rs
