+++ api/src/models/translations.rs
@@ -150,9 +150,10 @@ impl<'q> sqlx::Encode<'q, sqlx::Postgres> for TextContentId {
 
 // ======================================
 //
-// Translations in JSONB columns
+// * JSON TRANSLATIONS *
 //
-// Traits and helper functions for resolving translations in nested JSONB columns.
+// Traits and helper functions for resolving translations in nested JSON
+// structures (i.e. JSONB columns on database rows.)
 //
 // ======================================
 
@@ -188,27 +189,27 @@ pub trait CollectTextContentIds {
 }
 
 /// Implemented by types that contain [`TextContentId`] references and can
-/// be converted into a "resolved" mirror type with those references
+/// be converted into a "localized" mirror type with those references
 /// replaced by their localized `String` content.
 ///
 /// Implemented directly for [`TextContentId`] (the base case, which looks
 /// itself up in `map`), with blanket implementations for `Option<T>` and
 /// `Vec<T>` so composite types resolve automatically through those
-/// wrappers. Structs and enums generated by `#[derive(Translatable)]`
+/// wrappers. Structs and enums generated by `#[derive(TranslatableJson)]`
 /// implement this by recursing into each `#[translatable]` field and
 /// passing non-translatable fields through unchanged.
 ///
 /// Resolution is intentionally infallible: a [`TextContentId`] missing from
-/// `map` (see [`resolve_translations`] for when this can happen) resolves
+/// `map` (see [`localize_translations`] for when this can happen) resolves
 /// to an empty string rather than propagating an error, so a single missing
 /// translation can't fail an entire response.
-pub trait ResolveTranslations {
+pub trait LocalizeTranslations {
     /// The resolved mirror of `Self`, with every [`TextContentId`] replaced
     /// by its resolved `String`. For [`TextContentId`] itself this is
     /// `String`; for a derived struct/enum, this is its generated
-    /// `*Resolved` counterpart; for `Option<T>`/`Vec<T>`, this is
-    /// `Option<T::Resolved>`/`Vec<T::Resolved>`.
-    type Resolved;
+    /// `*Localized` counterpart; for `Option<T>`/`Vec<T>`, this is
+    /// `Option<T::Localized>`/`Vec<T::Localized>`.
+    type Localized;
 
     /// Consumes `self`, replacing every contained [`TextContentId`] with
     /// its resolved text from `map`.
@@ -216,23 +217,55 @@ pub trait ResolveTranslations {
     /// # Arguments
     ///
     /// * `map` - Pre-fetched translations, as returned by
-    ///   [`resolve_translations`]. Must have been built from a superset of
+    ///   [`localize_translations`]. Must have been built from a superset of
     ///   the ids [`CollectTextContentIds::collect_text_content_ids`] would
     ///   gather from `self`, or missing ids can silently resolve to an
     ///   empty string.
-    fn resolve(self, map: &HashMap<TextContentId, String>) -> Self::Resolved;
+    fn localize(self, map: &HashMap<TextContentId, String>) -> Self::Localized;
 }
 
+/// The resolved display text for a specific locale, alongside the full
+/// multi-locale translation record it was resolved from.
+///
+/// This is the admin-facing counterpart to the plain `String` produced by
+/// [`LocalizeTranslations`]. Where that path is for read-only display (a single
+/// localized string), this shape can be consumed with the existing translation
+/// admin UI.
+///
+/// Serialized with `camelCase` field names for frontend consumption.
 #[derive(Serialize, Deserialize, Debug, Clone, JsonSchema, PartialEq)]
 #[serde(rename_all = "camelCase")]
 pub struct JsonFieldWithTranslations {
+    /// The resolved text for the locale requested when this value was
+    /// built (see [`ResolveWithTranslations::resolve_with_translations`]),
+    /// falling back to the content's primary locale per [`pick_localized`].
     pub localized: String,
+    /// The full translation record (`text_content` row plus every
+    /// `text_translation` row) this value was resolved from.
     pub translations: TranslationDto,
 }
 
+/// Implemented by types that contain [`TextContentId`] references and can
+/// be converted into a mirror type where those references are replaced by
+/// [`JsonFieldWithTranslations`] — the resolved text *plus* its full
+/// translation record — rather than a bare resolved `String`.
+///
+/// This is a sibling to [`LocalizeTranslations`], not a replacement: use
+/// `LocalizeTranslations` for ordinary read-only display and
+/// `ResolveWithTranslations` for admin/editing contexts.
+///
+/// Implemented directly for [`TextContentId`] with blanket implementations
+/// expected for `Option<T>`/`Vec<T>` so composite types resolve through those
+/// wrappers.
 pub trait ResolveWithTranslations {
+    /// The mirror of `Self` with every [`TextContentId`] replaced by
+    /// [`JsonFieldWithTranslations`]. For [`TextContentId`] itself this is
+    /// `JsonFieldWithTranslations`; for a derived struct/enum, this is the
+    /// analogous "with translations" generated counterpart.
     type WithTranslations;
 
+    /// Consumes `self`, replacing every contained [`TextContentId`] with a
+    /// [`JsonFieldWithTranslations`] built from `translations` and `locale`.
     fn resolve_with_translations(
         self,
         translations: &HashMap<TextContentId, TranslationDto>,
@@ -275,32 +308,6 @@ impl ResolveWithTranslations for TextContentId {
                 }
             }
         }
-    }
-}
-
-impl<T: ResolveWithTranslations> ResolveWithTranslations for Option<T> {
-    type WithTranslations = Option<T::WithTranslations>;
-
-    fn resolve_with_translations(
-        self,
-        translations: &HashMap<TextContentId, TranslationDto>,
-        locale: &str,
-    ) -> Self::WithTranslations {
-        self.map(|v| v.resolve_with_translations(translations, locale))
-    }
-}
-
-impl<T: ResolveWithTranslations> ResolveWithTranslations for Vec<T> {
-    type WithTranslations = Vec<T::WithTranslations>;
-
-    fn resolve_with_translations(
-        self,
-        translations: &HashMap<TextContentId, TranslationDto>,
-        locale: &str,
-    ) -> Self::WithTranslations {
-        self.into_iter()
-            .map(|v| v.resolve_with_translations(translations, locale))
-            .collect()
     }
 }
 
@@ -376,11 +383,11 @@ impl<T: BuildTextTranslation + Send> BuildTextTranslation for Option<T> {
     }
 }
 
-// ======================================
+// =============================================
 //
-// Implementations for [`TextContentId`]
+// Blanket implementations for [`TextContentId`]
 //
-// ======================================
+// =============================================
 
 impl CollectTextContentIds for TextContentId {
     fn collect_text_content_ids(&self, out: &mut HashSet<TextContentId>) {
@@ -404,27 +411,53 @@ impl<T: CollectTextContentIds> CollectTextContentIds for Vec<T> {
     }
 }
 
-impl ResolveTranslations for TextContentId {
-    type Resolved = String;
+impl LocalizeTranslations for TextContentId {
+    type Localized = String;
 
-    fn resolve(self, map: &HashMap<TextContentId, String>) -> Self::Resolved {
+    fn localize(self, map: &HashMap<TextContentId, String>) -> Self::Localized {
         map.get(&self).cloned().unwrap_or_default()
     }
 }
 
-impl<T: ResolveTranslations> ResolveTranslations for Option<T> {
-    type Resolved = Option<T::Resolved>;
+impl<T: LocalizeTranslations> LocalizeTranslations for Option<T> {
+    type Localized = Option<T::Localized>;
 
-    fn resolve(self, map: &HashMap<TextContentId, String>) -> Self::Resolved {
-        self.map(|v| v.resolve(map))
+    fn localize(self, map: &HashMap<TextContentId, String>) -> Self::Localized {
+        self.map(|v| v.localize(map))
     }
 }
 
-impl<T: ResolveTranslations> ResolveTranslations for Vec<T> {
-    type Resolved = Vec<T::Resolved>;
+impl<T: LocalizeTranslations> LocalizeTranslations for Vec<T> {
+    type Localized = Vec<T::Localized>;
 
-    fn resolve(self, map: &HashMap<TextContentId, String>) -> Self::Resolved {
-        self.into_iter().map(|v| v.resolve(map)).collect()
+    fn localize(self, map: &HashMap<TextContentId, String>) -> Self::Localized {
+        self.into_iter().map(|v| v.localize(map)).collect()
+    }
+}
+
+impl<T: ResolveWithTranslations> ResolveWithTranslations for Option<T> {
+    type WithTranslations = Option<T::WithTranslations>;
+
+    fn resolve_with_translations(
+        self,
+        translations: &HashMap<TextContentId, TranslationDto>,
+        locale: &str,
+    ) -> Self::WithTranslations {
+        self.map(|v| v.resolve_with_translations(translations, locale))
+    }
+}
+
+impl<T: ResolveWithTranslations> ResolveWithTranslations for Vec<T> {
+    type WithTranslations = Vec<T::WithTranslations>;
+
+    fn resolve_with_translations(
+        self,
+        translations: &HashMap<TextContentId, TranslationDto>,
+        locale: &str,
+    ) -> Self::WithTranslations {
+        self.into_iter()
+            .map(|v| v.resolve_with_translations(translations, locale))
+            .collect()
     }
 }
 
@@ -446,13 +479,13 @@ impl<T: ResolveTranslations> ResolveTranslations for Vec<T> {
 ///
 /// # Returns
 ///
-/// A map from each resolvable [`TextContentId`] to its resolved text.
+/// A map from each resolvable [`TextContentId`] to its localized text.
 ///
 /// Ids that have neither a `locale` translation nor a `primary_locale`
 /// fallback translation (e.g. a `text_content` row with no associated
 /// `text_translation` rows at all — normally a data integrity issue) are
 /// silently omitted from the returned map. Callers resolving via
-/// [`ResolveTranslations::resolve`] will fall back to an empty string for
+/// [`LocalizeTranslations::localize`] will fall back to an empty string for
 /// any id missing from the map; callers needing to detect this case
 /// explicitly should diff the returned map's keys against `ids`.
 ///
@@ -462,7 +495,7 @@ impl<T: ResolveTranslations> ResolveTranslations for Vec<T> {
 /// # Errors
 ///
 /// Returns [`ComhairleError`] if the query fails to execute.
-pub async fn resolve_translations(
+pub async fn localize_translations(
     db: &PgPool,
     ids: &[TextContentId],
     locale: &str,
@@ -1390,10 +1423,10 @@ mod tests {
         }
     }
 
-    impl ResolveTranslations for TestWithTranslations {
-        type Resolved = LocalizedTestWithTranslations;
+    impl LocalizeTranslations for TestWithTranslations {
+        type Localized = LocalizedTestWithTranslations;
 
-        fn resolve(self, map: &HashMap<TextContentId, String>) -> Self::Resolved {
+        fn localize(self, map: &HashMap<TextContentId, String>) -> Self::Localized {
             LocalizedTestWithTranslations {
                 title: map.get(&self.title).cloned().unwrap_or_default(),
                 description: map.get(&self.description).cloned().unwrap_or_default(),
@@ -1411,7 +1444,7 @@ mod tests {
 
         let ids = [trans_a.id, trans_b.id, trans_c.id];
 
-        let result = resolve_translations(&pool, &ids, "en").await?;
+        let result = localize_translations(&pool, &ids, "en").await?;
 
         assert_eq!(
             result.get(&trans_a.id).unwrap(),
@@ -1442,7 +1475,7 @@ mod tests {
 
         let ids = [trans_a.id, trans_b.id, trans_c.id];
 
-        let result = resolve_translations(&pool, &ids, "fr").await?;
+        let result = localize_translations(&pool, &ids, "fr").await?;
 
         assert_eq!(result.get(&trans_a.id).unwrap(), "Test text a");
         assert_eq!(result.get(&trans_b.id).unwrap(), "Test text b");
@@ -1492,9 +1525,9 @@ mod tests {
 
         test_translation.collect_text_content_ids(&mut out);
 
-        let map = resolve_translations(&pool, &out.into_iter().collect::<Vec<_>>(), "en").await?;
+        let map = localize_translations(&pool, &out.into_iter().collect::<Vec<_>>(), "en").await?;
 
-        let resolved = test_translation.resolve(&map);
+        let resolved = test_translation.localize(&map);
 
         assert_eq!(resolved.title, "Test title", "incorrect title");
         assert_eq!(
@@ -1544,8 +1577,8 @@ mod tests {
 
         let mut out: HashSet<TextContentId> = HashSet::new();
         root.collect_text_content_ids(&mut out);
-        let map = resolve_translations(&pool, &out.into_iter().collect::<Vec<_>>(), "en").await?;
-        let resolved = root.resolve(&map);
+        let map = localize_translations(&pool, &out.into_iter().collect::<Vec<_>>(), "en").await?;
+        let resolved = root.localize(&map);
 
         assert_eq!(
             resolved.title,
